requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
djangorestframework-simplejwt==5.3.1
celery==5.4.0
drf-yasg==1.21.7
stripe==10.2.0